email失败重推

+++ intramirror-root/intramirror-order-service/intramirror-order-service-core/src/main/java/com/intramirror/order/core/utils/ShipMailSendThread.java
@@ -121,20 +121,24 @@ public class ShipMailSendThread implements Runnable {
             mailContent.setToEmails(MailConfig.emailTo);
         }
 
+        boolean flag = true; //是否删除附件文件
         try {
             logger.info("ShipMailSendThread 开始发送邮件 content={}", new Gson().toJson(mailContent));
             MailSendUtil.sendMail(mailContent);
-            logger.info("ShipMailSendThread 邮件发送完成");
+            logger.info("ShipMailSendThread 邮件发送完成，shipmentNo:{}",shipment.getShipmentNo());
         } catch (Exception e) {
+            flag = false;
             e.printStackTrace();
-            logger.error("ShipMailSendThread 邮件发送失败", e);
+            logger.error("ShipMailSendThread 邮件发送失败，shipmentNo:{},errorMsg:{}",shipment.getShipmentNo(), e);
         } finally {
-            //删除附件文件
-            List<MailAttachmentVO> attachments = mailContent.getAttachments();
-            for (MailAttachmentVO attachment : attachments) {
-                File file = new File(attachment.getFileUrl());
-                if (file.exists() && file.isFile()) {
-                    file.delete();
+            if (flag){
+                //删除附件文件
+                List<MailAttachmentVO> attachments = mailContent.getAttachments();
+                for (MailAttachmentVO attachment : attachments) {
+                    File file = new File(attachment.getFileUrl());
+                    if (file.exists() && file.isFile()) {
+                        file.delete();
+                    }
                 }
             }
         }
